refactor: share one API Gateway principal
REST APIs and HTTP APIs invoke a Lambda function as the same service
principal, and each service declared its own constant for it. The
apigatewayv2 copy is gone and the v2 serving layer reads the one at
`src/service/apigateway/sim-api-gateway-service-principal.ts`, which
follows the `sim-<service>-service-principal.ts` modules CloudFront and S3
use. The two files in `apigatewayv2/api/` that imported the constant from
`apigatewayv2/serve/auth/` now read it from there too, so the API model no
longer reaches up into the serving layer that evaluates the grant.
